@@ -5,7 +5,7 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=5.6.4",
+        "php": ">=7.1.3",
         "barryvdh/laravel-dompdf": "^0.9.0",
         "bepsvpt/secure-headers": "^6.0",
         "biscolab/laravel-recaptcha": "^4.1",
@@ -18,14 +18,14 @@
         "laravelcollective/html": "~5.0",
         "maatwebsite/excel": "^3.1",
         "orchestra/parser": "^3.5",
-        "phpoffice/phpspreadsheet": "*",
-        "pusher/pusher-php-server": "*",
-        "smalot/pdfparser": "*"
+        "phpoffice/phpspreadsheet": "^1.23",
+        "pusher/pusher-php-server": "^5.0",
+        "smalot/pdfparser": "^0.16"
     },
     "require-dev": {
         "fzaninotto/faker": "~1.4",
-        "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~5.7"
+        "mockery/mockery": "^1.3",
+        "phpunit/phpunit": "~7.0"
     },
     "autoload": {
 		"files": [
