@@ -5,9 +5,11 @@ namespace App;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Roles extends Model
+class Hego extends Model
 {
-    use SoftDeletes;
-    
-	protected $table = 'roles';
+	use SoftDeletes;
+
+	protected $fillable = [
+		'nombre'
+	];
 }
