@@ -58,6 +58,8 @@ use App\DatamesSemZipaquira;
 use App\DatamesGen;
 use App\Pagadurias;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
 class PagaduriasController extends Controller
@@ -163,7 +165,20 @@ class PagaduriasController extends Controller
 
     public function perDoc($doc)
     {
-        Log::info("ENtro a buscar la cedula: ",  ['doc' => $doc]);
+        Log::info("Entro a buscar la cedula: ",  ['doc' => $doc]);
+
+        $user = Auth::user();
+        $userType = IsCompany() ? 'empresa' : (IsComercial() ? 'comercial' : null);
+
+        if ($userType) {
+            if ($user->$userType->consultas_diarias <= 0) {
+                return response()->json([
+                    'message' => 'No tienes consultas disponibles',
+                ], 400);
+            }
+        }
+
+        DB::beginTransaction();
 
         try {
             $models = [
@@ -224,9 +239,9 @@ class PagaduriasController extends Controller
 
             foreach ($models as $model => $column) {
                 $data = $model::where($column, 'LIKE', '%' . $doc . '%')
-                ->orderBy('id', 'desc')
-                ->first();
- 
+                    ->orderBy('id', 'desc')
+                    ->first();
+
                 if ($data) {
                     $modelName = class_basename($model);
                     $results[Str::camel($modelName)] = $data;
@@ -234,7 +249,8 @@ class PagaduriasController extends Controller
             }
 
             // $dataGen = DatamesGen::where('doc', 'like', '%' . $doc . '%')->get();
-            $dataGen = DatamesGen::where('doc', '=', $doc )->get();
+            $dataGen = DatamesGen::where('doc', '=', $doc)->get();
+
             if ($dataGen) {
                 foreach ($dataGen as $item) {
                     // Verifica si la pagaduría ya existe en $results
@@ -249,15 +265,23 @@ class PagaduriasController extends Controller
                 }
             }
 
+            if ($userType) {
+                $user->$userType->decrement('consultas_diarias', 1);
+            }
+
+            DB::commit();
+
             $results = !empty($results) ? $results : (object) [];
 
             return response()->json($results, 200);
         } catch (\Exception $e) {
+            DB::rollBack();
+
             Log::error("Error al buscar pagadurías por documento: {$e->getMessage()}", [
                 'doc' => $doc, // Incluye el documento para identificar la consulta que causó el error
                 'exception' => $e->getTraceAsString(), // Para un diagnóstico más detallado
             ]);
-        
+
             return response()->json([
                 'error' => 'Ocurrió un error al procesar la solicitud',
                 'message' => $e->getMessage(), // Mensaje exacto del error
@@ -266,104 +290,172 @@ class PagaduriasController extends Controller
     }
 
     public function getPagaduriasNames()
-{
-    $nombres = Pagadurias::pluck('pagaduria');
-    return response()->json($nombres, 200);
-}
-
-public function getPagaduriasNamesAmi()
-{
-    $nombres = [
-        "SED AMAZONAS", "SED ANTIOQUIA", "SED ARAUCA", "SED ATLANTICO",
-        "SED BOLIVAR", "SED BOYACA", "SED CALDAS", "SED CAQUETA",
-        "SED CASANARE", "SED CAUCA", "SED CESAR", "SED CHOCO",
-        "SED CORDOBA", "SED CUNDINAMARCA", "SED GUAJIRA", "SED GUAVIARE",
-        "SED HUILA", "SED MAGDALENA", "SED META", "SED NARINO",
-        "SED NORTE DE SANTANDER", "SED PUTUMAYO", "SED QUINDIO", "SED RISARALDA",
-        "SED SANTANDER", "SED SINCELEJO", "SED SUCRE", "SED TOLIMA",
-        "SED VALLE", "SED VAUPES", "SED VICHADA", "SEM APARTADO",
-        "SEM ARMENIA", "SEM BARRANCABERMEJA", "SEM BARRANQUILLA", "SEM BELLO",
-        "SEM BUCARAMANGA", "SEM BUENAVENTURA", "SEM BUGA", "SEM CALI",
-        "SEM CARTAGENA", "SEM CARTAGO", "SEM CHIA", "SEM CIENAGA",
-        "SEM CUCUTA", "SEM DOSQUEBRADAS", "SEM DUITAMA", "SEM ENVIGADO",
-        "SEM FACATATIVA", "SEM FLORENCIA", "SEM FLORIDABLANCA", "SEM FUNZA",
-        "SEM FUSAGASUGA", "SEM GIRARDOT", "SEM GIRON", "SEM GUAINIA",
-        "SEM IBAGUE", "SEM IPIALES", "SEM ITAGUI", "SEM JAMUNDI",
-        "SEM LORICA", "SEM MAGANGUE", "SEM MAICAO", "SEM MALAMBO",
-        "SEM MANIZALES", "SEM MEDELLIN", "SEM MONTERIA", "SEM MOSQUERA",
-        "SEM NEIVA", "SEM PALMIRA", "SEM PASTO", "SEM PEREIRA",
-        "SEM PIEDECUESTA", "SEM PITALITO", "SEM POPAYAN", "SEM QUIBDO",
-        "SEM RIOHACHA", "SEM RIONEGRO", "SEM SABANETA", "SEM SAHAGUN",
-        "SEM SAN ANDRES", "SEM SANTA MARTA", "SEM SOACHA", "SEM SOGAMOSO",
-        "SEM SOLEDAD", "SEM TULUA", "SEM TUMACO", "SEM TUNJA",
-        "SEM TURBO", "SEM URIBIA", "SEM VALLEDUPAR", "SEM VILLAVICENCIO",
-        "SEM YOPAL", "SEM YUMBO", "SEM ZIPAQUIRA"
-    ];
-    
-
-    return response()->json($nombres, 200);
-}
-
-public function getSituacionLaboralByDoc($doc)
-{
-    try {
-        Log::info("Consultando la situación laboral para el documento: {$doc}");
-
-        $data = DatamesGen::where('doc', 'like', "%{$doc}%")
-                          ->latest('id')
-                          ->first(['situacion_laboral']);
-
-        if (!$data) {
-            Log::info("No se encontró la situación laboral para el documento proporcionado.", ['doc' => $doc]);
-            return response()->json(['mensaje' => 'No se encontró la situación laboral para el documento proporcionado.'], 404);
-        }
-
-        Log::info("Situación laboral encontrada.", ['doc' => $doc, 'situacion_laboral' => $data->situacion_laboral]);
-
-        return response()->json($data->situacion_laboral, 200);
-
-    } catch (\Exception $e) {
-        Log::error("Error al buscar la situación laboral por documento: {$doc}", [
-            'doc' => $doc,
-            'exception' => $e->getMessage(),
-            'stack' => $e->getTraceAsString(),
-        ]);
-
-        return response()->json(['error' => 'Ocurrió un error al procesar la solicitud'], 500);
+    {
+        $nombres = Pagadurias::pluck('pagaduria');
+        return response()->json($nombres, 200);
     }
-}
 
-public function getSituacionLaboralByDocs(Request $request)
-{
-    try {
-        $documentos = $request->input('documentos', []);
-        Log::info("Consultando la situación laboral para múltiples documentos.");
+    public function getPagaduriasNamesAmi()
+    {
+        $nombres = [
+            "SED AMAZONAS",
+            "SED ANTIOQUIA",
+            "SED ARAUCA",
+            "SED ATLANTICO",
+            "SED BOLIVAR",
+            "SED BOYACA",
+            "SED CALDAS",
+            "SED CAQUETA",
+            "SED CASANARE",
+            "SED CAUCA",
+            "SED CESAR",
+            "SED CHOCO",
+            "SED CORDOBA",
+            "SED CUNDINAMARCA",
+            "SED GUAJIRA",
+            "SED GUAVIARE",
+            "SED HUILA",
+            "SED MAGDALENA",
+            "SED META",
+            "SED NARINO",
+            "SED NORTE DE SANTANDER",
+            "SED PUTUMAYO",
+            "SED QUINDIO",
+            "SED RISARALDA",
+            "SED SANTANDER",
+            "SED SINCELEJO",
+            "SED SUCRE",
+            "SED TOLIMA",
+            "SED VALLE",
+            "SED VAUPES",
+            "SED VICHADA",
+            "SEM APARTADO",
+            "SEM ARMENIA",
+            "SEM BARRANCABERMEJA",
+            "SEM BARRANQUILLA",
+            "SEM BELLO",
+            "SEM BUCARAMANGA",
+            "SEM BUENAVENTURA",
+            "SEM BUGA",
+            "SEM CALI",
+            "SEM CARTAGENA",
+            "SEM CARTAGO",
+            "SEM CHIA",
+            "SEM CIENAGA",
+            "SEM CUCUTA",
+            "SEM DOSQUEBRADAS",
+            "SEM DUITAMA",
+            "SEM ENVIGADO",
+            "SEM FACATATIVA",
+            "SEM FLORENCIA",
+            "SEM FLORIDABLANCA",
+            "SEM FUNZA",
+            "SEM FUSAGASUGA",
+            "SEM GIRARDOT",
+            "SEM GIRON",
+            "SEM GUAINIA",
+            "SEM IBAGUE",
+            "SEM IPIALES",
+            "SEM ITAGUI",
+            "SEM JAMUNDI",
+            "SEM LORICA",
+            "SEM MAGANGUE",
+            "SEM MAICAO",
+            "SEM MALAMBO",
+            "SEM MANIZALES",
+            "SEM MEDELLIN",
+            "SEM MONTERIA",
+            "SEM MOSQUERA",
+            "SEM NEIVA",
+            "SEM PALMIRA",
+            "SEM PASTO",
+            "SEM PEREIRA",
+            "SEM PIEDECUESTA",
+            "SEM PITALITO",
+            "SEM POPAYAN",
+            "SEM QUIBDO",
+            "SEM RIOHACHA",
+            "SEM RIONEGRO",
+            "SEM SABANETA",
+            "SEM SAHAGUN",
+            "SEM SAN ANDRES",
+            "SEM SANTA MARTA",
+            "SEM SOACHA",
+            "SEM SOGAMOSO",
+            "SEM SOLEDAD",
+            "SEM TULUA",
+            "SEM TUMACO",
+            "SEM TUNJA",
+            "SEM TURBO",
+            "SEM URIBIA",
+            "SEM VALLEDUPAR",
+            "SEM VILLAVICENCIO",
+            "SEM YOPAL",
+            "SEM YUMBO",
+            "SEM ZIPAQUIRA"
+        ];
 
-        $situaciones = DatamesGen::whereIn('doc', $documentos)
-                                ->latest('id')
-                                ->get()
-                                ->keyBy('doc')  // Esto asume que 'doc' es único
-                                ->map(function ($item) {
-                                    return $item->situacion_laboral;
-                                });
-
-        if ($situaciones->isEmpty()) {
-            Log::info("No se encontraron situaciones laborales para los documentos proporcionados.");
-            return response()->json(['mensaje' => 'No se encontraron situaciones laborales para los documentos proporcionados.'], 404);
-        }
-
-        Log::info("Situaciones laborales encontradas.", ['situaciones' => $situaciones]);
-
-        return response()->json($situaciones, 200);
-
-    } catch (\Exception $e) {
-        Log::error("Error al buscar las situaciones laborales por documentos", [
-            'documentos' => $documentos,
-            'exception' => $e->getMessage(),
-            'stack' => $e->getTraceAsString(),
-        ]);
-
-        return response()->json(['error' => 'Ocurrió un error al procesar la solicitud'], 500);
+        return response()->json($nombres, 200);
     }
-}
+
+    public function getSituacionLaboralByDoc($doc)
+    {
+        try {
+            Log::info("Consultando la situación laboral para el documento: {$doc}");
+
+            $data = DatamesGen::where('doc', 'like', "%{$doc}%")
+                ->latest('id')
+                ->first(['situacion_laboral']);
+
+            if (!$data) {
+                Log::info("No se encontró la situación laboral para el documento proporcionado.", ['doc' => $doc]);
+                return response()->json(['mensaje' => 'No se encontró la situación laboral para el documento proporcionado.'], 404);
+            }
+
+            Log::info("Situación laboral encontrada.", ['doc' => $doc, 'situacion_laboral' => $data->situacion_laboral]);
+
+            return response()->json($data->situacion_laboral, 200);
+        } catch (\Exception $e) {
+            Log::error("Error al buscar la situación laboral por documento: {$doc}", [
+                'doc' => $doc,
+                'exception' => $e->getMessage(),
+                'stack' => $e->getTraceAsString(),
+            ]);
+
+            return response()->json(['error' => 'Ocurrió un error al procesar la solicitud'], 500);
+        }
+    }
+
+    public function getSituacionLaboralByDocs(Request $request)
+    {
+        try {
+            $documentos = $request->input('documentos', []);
+            Log::info("Consultando la situación laboral para múltiples documentos.");
+
+            $situaciones = DatamesGen::whereIn('doc', $documentos)
+                ->latest('id')
+                ->get()
+                ->keyBy('doc')  // Esto asume que 'doc' es único
+                ->map(function ($item) {
+                    return $item->situacion_laboral;
+                });
+
+            if ($situaciones->isEmpty()) {
+                Log::info("No se encontraron situaciones laborales para los documentos proporcionados.");
+                return response()->json(['mensaje' => 'No se encontraron situaciones laborales para los documentos proporcionados.'], 404);
+            }
+
+            Log::info("Situaciones laborales encontradas.", ['situaciones' => $situaciones]);
+
+            return response()->json($situaciones, 200);
+        } catch (\Exception $e) {
+            Log::error("Error al buscar las situaciones laborales por documentos", [
+                'documentos' => $documentos,
+                'exception' => $e->getMessage(),
+                'stack' => $e->getTraceAsString(),
+            ]);
+
+            return response()->json(['error' => 'Ocurrió un error al procesar la solicitud'], 500);
+        }
+    }
 }
