@@ -82,6 +82,8 @@ Vue.component('CarteraIcon', require('./components/icons/CarteraIcon').default);
 Vue.component('InvestigacionIcon', require('./components/icons/InvestigacionIcon').default);
 Vue.component('LocalizacionIcon', require('./components/icons/LocalizacionIcon').default);
 Vue.component('DownloadIcon', require('./components/icons/DownloadIcon').default);
+Vue.component('UsersIcon', require('./components/icons/UsersIcon').default);
+Vue.component('DashboardIcon', require('./components/icons/DashboardIcon').default);
 
 //Intragration Pages
 Vue.component('integration', require('./components/pages/Integrations/index.vue').default);
@@ -95,8 +97,8 @@ Vue.component('certificados', require('./components/pages/ConsultDataClient/cert
 Vue.component('history-component', require('./components/pages/ConsultDataClient/history.vue').default);
 Vue.component('detail-history-component', require('./components/pages/ConsultDataClient/detailhistory.vue').default);
 Vue.component(
-    'detail-history-component-draft',
-    require('./components/pages/ConsultDataClient/detailhistoryBorrador.vue').default
+	'detail-history-component-draft',
+	require('./components/pages/ConsultDataClient/detailhistoryBorrador.vue').default
 );
 Vue.component('credit-calculator', require('./components/pages/CreditColsult/CreditCalculator.vue').default);
 Vue.component('register-credit', require('./components/pages/CreditColsult/CreditForm.vue').default);
@@ -105,16 +107,16 @@ Vue.component('CreditRequestsList', require('./components/pages/CreditRequest/Cr
 //AMI Integración en solicitud credito
 
 Vue.component(
-    'form-consult-integration',
-    require('./components/pages/ConsultDataClientDraft/FormConsultIntegration.vue').default
+	'form-consult-integration',
+	require('./components/pages/ConsultDataClientDraft/FormConsultIntegration.vue').default
 );
 Vue.component(
-    'consulta-pagadurias-form-consult',
-    require('./components/pages/ConsultDataClientDraft/ConsultaPagaduriasFormConsult.vue').default
+	'consulta-pagadurias-form-consult',
+	require('./components/pages/ConsultDataClientDraft/ConsultaPagaduriasFormConsult.vue').default
 );
 Vue.component(
-    'client-data-component-draft-integration',
-    require('./components/pages/ConsultDataClientDraft/indexIntegration.vue').default
+	'client-data-component-draft-integration',
+	require('./components/pages/ConsultDataClientDraft/indexIntegration.vue').default
 );
 
 //CUPONES
@@ -130,20 +132,22 @@ Vue.component('UploadPensiones', UploadPensiones);
 Vue.component('UploadFiducidiaria', UploadFiducidiaria);
 Vue.component('join-pensiones', JoinPensiones);
 
-// Registro de CreditForm
-Vue.component(
-    'credit-form',
-    require('./components/pages/CreditRequest/CreditForm.vue').default
-  );
+Vue.component('empresas', require('./components/pages/Empresas/Index.vue').default);
+Vue.component('crear-empresas', require('./components/pages/Empresas/Create.vue').default);
+Vue.component('editar-empresas', require('./components/pages/Empresas/Edit.vue').default);
+Vue.component('ver-empresas', require('./components/pages/Empresas/Ver.vue').default);
 
-  Vue.component('credit-requests-list',
-    require('./components/pages/ConsultDataClientDraft/indexCreditRequest.vue').default
-);
+Vue.component('area-comerciales', require('./components/pages/AreaComerciales/Index.vue').default);
+Vue.component('crear-area-comerciales', require('./components/pages/AreaComerciales/Create.vue').default);
+Vue.component('editar-area-comerciales', require('./components/pages/AreaComerciales/Edit.vue').default);
+Vue.component('ver-area-comerciales', require('./components/pages/AreaComerciales/Ver.vue').default);
 
-  
+Vue.component('sedes', require('./components/pages/Sedes/Index.vue').default);
+Vue.component('crear-sedes', require('./components/pages/Sedes/Create.vue').default);
+Vue.component('editar-sedes', require('./components/pages/Sedes/Edit.vue').default);
 
 /* Vue Init */
 const app = new Vue({
-    el: '#app',
-    store
+	el: '#app',
+	store
 });
