@@ -202,12 +202,12 @@
                     <div
                         class="col-6"
                         v-if="
-                            user.roles_id === 1 ||
-                            user.roles_id === '1' ||
-                            user.roles_id === 4 ||
-                            user.roles_id === '4' ||
-                            user.roles_id === 5 ||
-                            user.roles_id === '5'
+                            user.role_id === 1 ||
+                            user.role_id === '1' ||
+                            user.role_id === 4 ||
+                            user.role_id === '4' ||
+                            user.role_id === 5 ||
+                            user.role_id === '5'
                         "
                     >
                         <b class="panel-label">FECHA CARGA DATA:</b>
@@ -218,12 +218,12 @@
                     <div
                         class="col-6"
                         v-if="
-                            user.roles_id === 1 ||
-                            user.roles_id === '1' ||
-                            user.roles_id === 4 ||
-                            user.roles_id === '4' ||
-                            user.roles_id === 5 ||
-                            user.roles_id === '5'
+                            user.role_id === 1 ||
+                            user.role_id === '1' ||
+                            user.role_id === 4 ||
+                            user.role_id === '4' ||
+                            user.role_id === 5 ||
+                            user.role_id === '5'
                         "
                     >
                         <b class="panel-label">MES CARGA DATA:</b>
@@ -234,12 +234,12 @@
                     <div
                         class="col-6"
                         v-if="
-                            user.roles_id === 1 ||
-                            user.roles_id === '1' ||
-                            user.roles_id === 4 ||
-                            user.roles_id === '4' ||
-                            user.roles_id === 5 ||
-                            user.roles_id === '5'
+                            user.role_id === 1 ||
+                            user.role_id === '1' ||
+                            user.role_id === 4 ||
+                            user.role_id === '4' ||
+                            user.role_id === 5 ||
+                            user.role_id === '5'
                         "
                     >
                         <b class="panel-label">AÑO CARGA DATA:</b>
