@@ -75,13 +75,11 @@
                                 <table role="table" aria-colcount="1" class="table b-table table-striped table-hover">
 
                                     <thead role="rowgroup" class="table-header-nowrap">
-              
                                         <tr role="row">
                                             <th role="columnheader" scope="col" aria-colindex="1">Fecha de Ingreso</th>
                                         </tr>
                                     </thead>
                                     <tbody role="rowgroup">
-                                       
                                         <tr role="row" v-for="(item, index) in datamesSedArray" :key="index">
                                             <td aria-colindex="1" role="cell">{{ item.fecha_ingreso || '--' }}</td>
                                         </tr>
@@ -649,12 +647,12 @@
                 <div
                     class="col-6"
                     v-if="
-                        user.roles_id === 1 ||
-                        user.roles_id === '1' ||
-                        user.roles_id === 4 ||
-                        user.roles_id === '4' ||
-                        user.roles_id === 5 ||
-                        user.roles_id === '5'
+                        user.role_id === 1 ||
+                        user.role_id === '1' ||
+                        user.role_id === 4 ||
+                        user.role_id === '4' ||
+                        user.role_id === 5 ||
+                        user.role_id === '5'
                     "
                 >
                     <b class="panel-label">FECHA CARGA DATA:</b>
@@ -665,12 +663,12 @@
                 <div
                     class="col-6"
                     v-if="
-                        user.roles_id === 1 ||
-                        user.roles_id === '1' ||
-                        user.roles_id === 4 ||
-                        user.roles_id === '4' ||
-                        user.roles_id === 5 ||
-                        user.roles_id === '5'
+                        user.role_id === 1 ||
+                        user.role_id === '1' ||
+                        user.role_id === 4 ||
+                        user.role_id === '4' ||
+                        user.role_id === 5 ||
+                        user.role_id === '5'
                     "
                 >
                     <b class="panel-label">MES CARGA DATA:</b>
@@ -681,12 +679,12 @@
                 <div
                     class="col-6"
                     v-if="
-                        user.roles_id === 1 ||
-                        user.roles_id === '1' ||
-                        user.roles_id === 4 ||
-                        user.roles_id === '4' ||
-                        user.roles_id === 5 ||
-                        user.roles_id === '5'
+                        user.role_id === 1 ||
+                        user.role_id === '1' ||
+                        user.role_id === 4 ||
+                        user.role_id === '4' ||
+                        user.role_id === 5 ||
+                        user.role_id === '5'
                     "
                 >
                     <b class="panel-label">AÑO CARGA DATA:</b>
@@ -812,31 +810,38 @@ export default {
     }
 };
 </script>
+
 <style scoped>
 th {
     font-size: 14px;
 }
+
 ul {
     margin: 0;
     padding: 0;
 }
+
 li {
     list-style-type: none;
     margin-bottom: 10px;
     padding: 0;
     border-bottom: none;
 }
+
 p {
     margin: 0;
     font-size: 14px;
 }
+
 b {
     font-size: 14px;
 }
+
 .table thead th {
     font-weight: 600;
     vertical-align: middle;
 }
+
 @media (min-width: 770px) {
     .col-48 {
         max-width: 48%;
