@@ -8,7 +8,7 @@
 
 		<!-- CSRF Token -->
 		<meta name="csrf-token" content="{{ csrf_token() }}">
-		<title>{{ Auth::user()->rol->rol == 'ADMIN_HEGO' ? 'HEGO' : 'AMI' }} - @yield('title')</title>
+		<title>{{ Auth::user()->role->name == 'ADMIN_HEGO' ? 'HEGO' : 'AMI' }} - @yield('title')</title>
 
 		<base href="{{url('/')}}">
 
@@ -18,13 +18,13 @@
 		<link href="{{asset('css/simple-sidebar.css')}}" rel="stylesheet">
 		<link href="{{asset('css/font-awesome-4.7.0/css/font-awesome.min.css')}}" rel="stylesheet">
 		<link href="{{asset('css/styles.css')}}" rel="stylesheet">
-		
+
 		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/latest/css/toastr.min.css">
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>
-		<script src="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/2.0.2/js/toastr.min.js"></script>    
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/2.0.2/js/toastr.min.js"></script>
 		@yield('css')
 	</head>
-	<body class="skin-{{ Auth::user()->rol->rol == 'ADMIN_HEGO' ? 'red' : 'blue' }} sidebar-mini sidebar-open">
+	<body class="skin-{{ Auth::user()->role->name == 'ADMIN_HEGO' ? 'red' : 'blue' }} sidebar-mini sidebar-open">
 		<!-- Loader -->
 		<div class="loader">
 			<div class="text-center">
