@@ -19,39 +19,39 @@ use Illuminate\Http\Request;
 
 
 // Aliados
-Route::group(['prefix' => 'aliados'], function(){
+Route::group(['prefix' => 'aliados'], function () {
     Route::post('/show', 'JsonAliadosController@show');
     Route::post('/getFactor', 'JsonAliadosController@getFactor');
 });
 
 // Cargos
-Route::group(['prefix' => 'cargos'], function(){
+Route::group(['prefix' => 'cargos'], function () {
     Route::post('/xVinculacion', 'JsonCargosController@xVinculacion');
     Route::post('/xCargo', 'JsonCargosController@xCargo');
 });
 
 // Ciudades
-Route::group(['prefix' => 'ciudades'], function(){
+Route::group(['prefix' => 'ciudades'], function () {
     Route::post('/xDepto', 'JsonCiudadesController@xDepto');
 });
 
 // Oficinas
-Route::group(['prefix' => 'oficinas'], function(){
+Route::group(['prefix' => 'oficinas'], function () {
     Route::post('/xCiudad', 'JsonOficinasController@xCiudad');
 });
 
 // Comerciales
-Route::group(['prefix' => 'comerciales'], function(){
+Route::group(['prefix' => 'comerciales'], function () {
     Route::post('/infoCliente', 'JsonComercialesController@infoCliente');
 });
 
 // Parametros
-Route::group(['prefix' => 'parametros'], function(){
+Route::group(['prefix' => 'parametros'], function () {
     Route::post('/xVinculacion', 'JsonParametrosController@xVinculacion');
 });
 
 // Terecuperamos
-Route::group(['prefix' => 'terecuperamos'], function(){
+Route::group(['prefix' => 'terecuperamos'], function () {
     Route::post('/calcularDecision', 'JsonTerecuperamosController@calcularDecision');
     Route::post('/paso1', 'JsonTerecuperamosController@paso1');
     Route::post('/paso2', 'JsonTerecuperamosController@paso2');
@@ -62,35 +62,33 @@ Route::group(['prefix' => 'terecuperamos'], function(){
     Route::post('/paso7', 'JsonTerecuperamosController@paso7');
 });
 
-
 // Carteras
-Route::group(['prefix' => 'carteras'], function(){
+Route::group(['prefix' => 'carteras'], function () {
     Route::post('/total', 'JsonCarterasController@total');
     Route::post('/show', 'JsonCarterasController@show');
     Route::post('/delete', 'JsonCarterasController@destroy');
     Route::post('/compraCk', 'JsonCarterasController@compraCk');
 });
 
-
 // Reportes
-Route::group(['prefix' => 'reportes'], function(){
+Route::group(['prefix' => 'reportes'], function () {
     Route::post('/consultas', 'JsonReportesController@consultas');
 });
 
 // Clientes
-Route::group(['prefix' => 'clientes'], function(){
+Route::group(['prefix' => 'clientes'], function () {
     Route::post('/getPagaduriasXPeriodo', 'JsonClientesController@getPagaduriasXPeriodo');
     Route::post('/getRegistrosXPagaduriayPeriodo', 'JsonClientesController@getRegistrosXPagaduriayPeriodo');
     Route::post('/actualizarRegistro', 'JsonClientesController@actualizarRegistro');
 });
 
 //AMI
-Route::group(['prefix' => 'ami'], function(){
+Route::group(['prefix' => 'ami'], function () {
     Route::post('/getDesprendiblesXDocumento', 'JsonConsultasController@getDesprendiblesXDocumento');
 });
 
 // JELOU
-Route::group(['prefix' => 'jelou'], function(){
+Route::group(['prefix' => 'jelou'], function () {
     Route::post('/createBaseReporteBiometria', 'JelouController@createBaseReporteBiometria');
     Route::post('/createReporteBiometria', 'JelouController@createReporteBiometria');
     Route::post('/createInformacionUsuario', 'JelouController@createInformacionUsuario');
