@@ -12,20 +12,22 @@ const mix = require("laravel-mix");
  */
 
 mix
-  .js("resources/assets/js/app.js", "public/js")
-  .sass("resources/assets/sass/app.scss", "public/css")
-  .sass("resources/assets/sass/amipersonas.scss", "public/css/pages")
-  .sass("resources/assets/sass/dashboard.scss", "public/css/pages")
-  .sass("resources/assets/sass/hego.scss", "public/css/pages")
-  .sass("resources/assets/sass/hegoinformation.scss", "public/css/pages")
-  .sass("resources/assets/sass/home.scss", "public/css/pages")
-  .sass("resources/assets/sass/login.scss", "public/css/pages")
-  .sass("resources/assets/sass/moreinformation.scss", "public/css/pages")
-  .webpackConfig({
-    resolve: {
-        fallback: {
-            "fs": false
-        },
-      },
-    })
-  .vue();
+	.js("resources/assets/js/app.js", "public/js")
+	.sass("resources/assets/sass/app.scss", "public/css")
+	.sass("resources/assets/sass/amipersonas.scss", "public/css/pages")
+	.sass("resources/assets/sass/dashboard.scss", "public/css/pages")
+	.sass("resources/assets/sass/hego.scss", "public/css/pages")
+	.sass("resources/assets/sass/hegoinformation.scss", "public/css/pages")
+	.sass("resources/assets/sass/home.scss", "public/css/pages")
+	.sass("resources/assets/sass/login.scss", "public/css/pages")
+	.sass("resources/assets/sass/moreinformation.scss", "public/css/pages")
+	.webpackConfig({
+		resolve: {
+			fallback: {
+				"fs": false
+			},
+		},
+	})
+	.vue();
+
+mix.disableNotifications();
